- Split connection tests in multiple classes for ease of maintaining - Added tests for Connection::send - Split Connection::delete into Connection::ack and Connection::reject - Connection::reject can either delete or not the message depending on delete_on_rejection
